experiments/requirements.txt
requests==2.25.1
python-dotenv
beautifulsoup4